Viewer mode + ACL everywhere

+++ app/Http/Controllers/CharacterController.php
@@ -46,6 +46,8 @@ class CharacterController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Character::class);
+
         return view('characters.create');
     }
 
@@ -57,6 +59,8 @@ class CharacterController extends Controller
      */
     public function store(StoreCharacter $request)
     {
+        $this->authorize('create', Character::class);
+
         Character::create($request->all());
         if ($request->has('submit-new')) {
             return redirect()->route($this->view . '.create')
@@ -73,6 +77,7 @@ class CharacterController extends Controller
      */
     public function show(Character $character)
     {
+        $this->authorize('view', $character);
         return view('characters.show', compact('character'));
     }
 
@@ -84,6 +89,8 @@ class CharacterController extends Controller
      */
     public function edit(Character $character)
     {
+        $this->authorize('update', $character);
+
         return view('characters.edit', compact('character'));
     }
 
@@ -96,6 +103,8 @@ class CharacterController extends Controller
      */
     public function update(StoreCharacter $request, Character $character)
     {
+        $this->authorize('update', $character);
+
         $character->update($request->all());
         if ($request->has('submit-new')) {
             return redirect()->route($this->view . '.create')
@@ -112,6 +121,8 @@ class CharacterController extends Controller
      */
     public function destroy(Character $character)
     {
+        $this->authorize('delete', $character);
+
         $character->delete();
         return redirect()->route($this->view . '.index')->with('success', trans($this->view . '.destroy.success'));
     }
